+++ vector_firmware/extra_packages/vector_interfaces/CMakeLists.txt
@@ -6,8 +6,6 @@ find_package(rosidl_default_generators REQUIRED)
 rosidl_generate_interfaces(${PROJECT_NAME}
   msg/EncoderTicks.msg
   msg/MotorTicksSec.msg
-  srv/ExampleService.srv
-  action/ExampleAction.action
   DEPENDENCIES
 )
