+++ vector_firmware/platformio.ini
@@ -14,7 +14,8 @@ board = esp32doit-devkit-v1
 framework = arduino
 
 board_microros_transport = serial   ;serial/wifi
-board_microros_user_meta = colcon.meta
+board_microros_distro = humble
+board_microros_user_meta = new_colcon.meta
 monitor_speed = 115200
 
 lib_deps =
